Refactored TicketsList by mainly dividing filter grid into components and adding hook for getting url parameters (#8553)
