+++ src/components/page-layouts/MainContentLayout.tsx
@@ -1,4 +1,5 @@
-import { Typography } from '@mui/material';
+import { SxProps, Typography } from '@mui/material';
+import { visuallyHidden } from '@mui/utils';
 import { ReactNode } from 'react';
 import { HeadTitle } from '../HeadTitle';
 import { VerticalBox } from '../styled/Wrappers';
@@ -6,15 +7,23 @@ import { VerticalBox } from '../styled/Wrappers';
 interface MainContentLayoutProps {
   headTitle?: string;
   heading?: string;
+  hiddenHeading?: boolean;
+  sx?: SxProps;
   children?: ReactNode;
 }
 
-export const MainContentLayout = ({ headTitle, heading, children }: MainContentLayoutProps) => {
+export const MainContentLayout = ({
+  headTitle,
+  heading,
+  hiddenHeading = false,
+  sx,
+  children,
+}: MainContentLayoutProps) => {
   return (
-    <VerticalBox component="section" sx={{ gap: '1rem' }}>
+    <VerticalBox component="section" sx={{ gap: '1rem', ...sx }}>
       <HeadTitle>{headTitle}</HeadTitle>
       {heading && (
-        <Typography variant="h1" gutterBottom>
+        <Typography variant="h1" sx={hiddenHeading ? visuallyHidden : {}} gutterBottom>
           {heading}
         </Typography>
       )}
